Move jarjar itself to common project
