+++ common/build.gradle
@@ -21,6 +21,8 @@ dependencies {
     api "codechicken:DiffPatch:${project.diffpatch_version}"
     api "commons-codec:commons-codec:${project.commons_codec_version}"
     api "net.neoforged:EclipseLaunchConfigs:${project.eclipse_launch_configs_version}"
+    api "net.neoforged:JarJarMetadata:${project.jarjar_version}"
+    api "net.neoforged:JarJarSelector:${project.jarjar_version}"
 
     // IDE support
     api "gradle.plugin.org.jetbrains.gradle.plugin.idea-ext:gradle-idea-ext:${project.gradle_idea_extension_version}"
