+++ gradle.properties
@@ -26,7 +26,7 @@ ivy_artifact_version=2.5.1
 httpclient_version=4.5.13
 srgutils_version=0.4.13
 diffpatch_version=1.5.0.29
-jarjar_version=0.3.19
+jarjar_version=0.4.1
 jetbrains_annotations_version=23.0.0
 gradle_idea_extension_version=1.1.6
 groovy_dsl_improver_version=1.0.10
